% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_similarity_matrix.R
\name{aa_similarity_matrix}
\alias{aa_similarity_matrix}
\title{Triplet similarity}
\usage{
aa_similarity_matrix(genetic_code = "5")
}
\arguments{
\item{genetic_code}{Genetic code as defined in the \code{getGeneticCode} function
of the \code{Biostrings} package.}
}
\value{
A data frame with three columns: \code{triplets_1}, \code{triplets_2},
  and \code{value} (pairwise Hamming similarity in [0, 1]).
}
\description{
This function calculates the pairwise similarity between triplets.
}
\details{
There is no need to run this function each time. The results can be stored and
uploaded when needed.
}
\examples{
\donttest{
aa_sim <- aa_similarity_matrix(genetic_code = "5")
}
}
