% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_similarity_score.R
\name{aa_similarity_score}
\alias{aa_similarity_score}
\title{Score based on aa similarity}
\usage{
aa_similarity_score(
  folder_path = NULL,
  output_path = NULL,
  asv_taxonomy = NULL,
  aa_sim_mat = NULL,
  genetic_code = "5",
  FUN_sim = min,
  FUN_gra = min
)
}
\arguments{
\item{folder_path}{Path to the folder where the results are stored.}

\item{output_path}{Path to the results folder. Default to folder path.}

\item{asv_taxonomy}{A file with the taxonomy assigned to ASV. It must have at least
3 columns: family, genus and species.}

\item{aa_sim_mat}{Pairwise similarity between triplets calculated with the function
\code{aa_similarity_matrix}.}

\item{genetic_code}{Genetic code as defined in the \code{getGeneticCode} function
of the \code{Biostrings} package.}

\item{FUN_sim}{Function to calculate the similarity among query and reference
triplets. Default to \code{min}.}

\item{FUN_gra}{Function to calculate the grantham distance among query and reference
triplets. Default to \code{min}.}
}
\value{
A tibble with the same columns as \code{asv_taxonomy}, joined with
  \code{triplet_similarity}, \code{grantham_score}, and
  \code{phylogenetic_score} columns from the computed results.
}
\description{
This function calculates the string similarity, the grantham distance and the phylogenetic
conservation of amino acids against a reference database.
}
\details{
The triplet of an amino acid at a specific position that is not found at the same
position in a reference dataset is compared with triplets of the same reference database.
Three indices are calculated: string similarity, grantham distance and phylogenetic conservation.
String similarity is calculated with the \code{hamming} distance of the
\code{stringdist} package. The grantham distance is calculated with the package
\code{grantham}, while the phylogenetic conservation is calculated by comparing
how much an amino acid is conserved at a given position of the reference database.
Continuous results are transformed into classes: conservative, moderately conservative,
moderately radical and radical.
}
